01/09

+++ server/controllers/product.controller.js
@@ -1,39 +1,44 @@
 const Product = require("../models/product.model");
 
-exports.product_get = (req, res) => {
-    Product.find(function (err, arr) {
-        console.log('err', err)
-        res.json(arr)
-    });
+
+
+exports.product_get = async (req, res) => {
+    try {
+        console.log(req.user.userId);
+        const tasks = await Product.find({ owner: req.user.userId });
+        res.json(tasks);
+    } catch (e) {
+        res.status(500).json({ message: "something goes wrong" });
+
+    }
 };
-exports.create_update = (req, res) => {
-    const data = {
-        id: req.body.id,
-        input: req.body.input,
-        checked: req.body.checked,
-    };
-    Product.findOne({ _id: req.body.id }, (err, post) => {
-        if (post) {
-            Product.findByIdAndUpdate(req.body.id, data, { upsert: false }).then(
-                updated => {
-                    res.json(updated);
-                }
-            );
+
+exports.create_update = (req, res) => { 
+
+    Product.findOne({ _id: req.body.id, owner: req.user.userId }, (err, post) => {
+        if (post) { 
+            console.log(`@@@@@@@@@@@@${post}`) 
+            console.log(`@@@@@@@@@@@@${req.body.id}`)
+            post.updateOne({ _id: req.body.id, owner: req.user.userId }, { $set: { _id: req.body.id, input: req.body.input, checked: req.body.checked } }, { multi: true });
+
         } else {
-            Product.create(data).then(created => {
-                res.json(created);
+
+            const task = new Product({
+                _id: req.body.id,
+                input: req.body.input,
+                checked: req.body.checked,
+                owner: req.user.userId
             });
+            task.save();
         }
     });
 };
-exports.delete_one = (req, res) => {
+exports.delete_one = async (req, res) => {
 
-    Product.findByIdAndDelete(req.params.id).then(post => {
-        res.json({ message: "Your post was deleted!" });
-    });
+    await Product.findOneAndDelete({ _id: req.params.id, owner: req.user.userId });
 };
-exports.check_all = (req, res) => {
-    Product.updateMany({ checked: !req.body.checked }, { $set: { checked: req.body.checked } }, { multi: true }, (err, result) => {
+exports.check_all = async (req, res) => {
+    Product.updateMany({ checked: !req.body.checked, owner: req.user.userId }, { $set: { checked: req.body.checked } }, { multi: true }, (err, result) => {
 
         if (err) {
             res.status(500).send(err);
@@ -42,9 +47,9 @@ exports.check_all = (req, res) => {
         res.send(result)
     });
 };
-exports.delete_all = (req, res) => {
+exports.delete_all = async (req, res) => {
 
-    Product.deleteMany({ checked: true }, { multi: true }, (err, result) => {
+    Product.deleteMany({ checked: true, owner: req.user.userId }, { multi: true }, (err, result) => {
 
         if (err) {
             res.status(500).send(err);
